% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddConfig.R
\name{addModuleConfig}
\alias{addModuleConfig}
\title{addModuleConfig}
\usage{
addModuleConfig(config, moduleConfig)
}
\arguments{
\item{config}{The current config to append the extra module to}

\item{moduleConfig}{A module config to be added}
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the config specification for one or more modules
}
\details{
User specifies the settings to create a config for a module
}
