% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultAboutConfig}
\alias{createDefaultAboutConfig}
\title{createDefaultAboutConfig}
\usage{
createDefaultAboutConfig(resultDatabaseDetails, useKeyring = T)
}
\arguments{
\item{resultDatabaseDetails}{A list containing the result database schema (schema), the tablePrefix for the results and other optional settings}

\item{useKeyring}{whether to save the resultDatabaseDetails to a keyring or system environmental variable}
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the about config specification
}
\details{
User specifies the settings to create a default config for an about module
}
