% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultCohortGeneratorConfig}
\alias{createDefaultCohortGeneratorConfig}
\title{createDefaultCohortGeneratorConfig}
\usage{
createDefaultCohortGeneratorConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the cohort generator config specification
}
\details{
User specifies the settings to create a default config for a cohort generator module
}
