% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultCohortMethodConfig}
\alias{createDefaultCohortMethodConfig}
\title{createDefaultCohortMethodConfig}
\usage{
createDefaultCohortMethodConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the cohort method config specification (depreciated)
}
\details{
User specifies the settings to create a default config for a cohort method module
}
