% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultDatasourcesConfig}
\alias{createDefaultDatasourcesConfig}
\title{createDefaultDatasourcesConfig}
\usage{
createDefaultDatasourcesConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the datasources config specification
}
\details{
User specifies the settings to create a default config for a datasources module
}
