% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultEvidenceSynthesisConfig}
\alias{createDefaultEvidenceSynthesisConfig}
\title{createDefaultEvidenceSynthesisConfig}
\usage{
createDefaultEvidenceSynthesisConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the Evidence Synthesis config specification (depreciated)
}
\details{
User specifies the settings to create a default config for a Evidence Synthesis module
}
