% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultHomeConfig}
\alias{createDefaultHomeConfig}
\title{createDefaultHomeConfig}
\usage{
createDefaultHomeConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the about config specification
}
\details{
User specifies the settings to create a default config for an about module
}
