% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultMetaConfig}
\alias{createDefaultMetaConfig}
\title{createDefaultMetaConfig}
\usage{
createDefaultMetaConfig(
  resultDatabaseDetails = list(tablePrefix = "es_", cmTablePrefix = "cm_",
    sccsTablePrefix = "sccs_", cgTablePrefix = "cg_", schema = schema, databaseMetaData =
    "DATABASE_META_DATA"),
  useKeyring = T
)
}
\arguments{
\item{resultDatabaseDetails}{A list containing the result database schema (schema), the tablePrefix for the results and other optional settings}

\item{useKeyring}{whether to save the resultDatabaseDetails to a keyring or system environmental variable}
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the Meta, or EvidenceSynthesis, config specification
}
\details{
User specifies the settings to create a default config for a Meta, or EvidenceSynthesis, module
}
