% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultPhevaluatorConfig}
\alias{createDefaultPhevaluatorConfig}
\title{createDefaultPhevaluatorConfig}
\usage{
createDefaultPhevaluatorConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the phevaluator config specification
}
\details{
User specifies the settings to create a default config for a phevaluator module
}
