% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultPredictionConfig}
\alias{createDefaultPredictionConfig}
\title{createDefaultPredictionConfig}
\usage{
createDefaultPredictionConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the prediction config specification
}
\details{
User specifies the settings to create a default config for a prediction module
}
