% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultReportConfig}
\alias{createDefaultReportConfig}
\title{createDefaultReportConfig}
\usage{
createDefaultReportConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the report config specification
}
\details{
User specifies the settings to create a default config for a report module
}
