% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultSccsConfig}
\alias{createDefaultSccsConfig}
\title{createDefaultSccsConfig}
\usage{
createDefaultSccsConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the SCCS config specification (depreciated)
}
\details{
User specifies the settings to create a default config for a SCCS module
}
