% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddConfig.R
\name{initializeModuleConfig}
\alias{initializeModuleConfig}
\title{initializeModuleConfig}
\usage{
initializeModuleConfig()
}
\value{
An empty list
}
\description{
Creates an empty config
}
\details{
An empty config that can be used to add shiny module configs to
}
