% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewShiny.R
\name{viewShiny}
\alias{viewShiny}
\title{viewShiny}
\usage{
viewShiny(config, connection = NULL, connectionDetails = NUL)
}
\arguments{
\item{config}{The json with the app config}

\item{connection}{A connection to the results}

\item{connectionDetails}{A DatabaseConnector::connectionDetails connection to the results database}
}
\value{
The shiny app will open
}
\description{
Open the shiny app
}
\details{
User specifies the json config and connection
}
