% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultCohortDiagnosticsConfig}
\alias{createDefaultCohortDiagnosticsConfig}
\title{createDefaultCohortDiagnosticsConfig}
\usage{
createDefaultCohortDiagnosticsConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the characterization config specification
}
\details{
User specifies the settings to create a default config for a characterization module
}
