% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultEstimationConfig}
\alias{createDefaultEstimationConfig}
\title{createDefaultEstimationConfig}
\usage{
createDefaultEstimationConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the estimation config specification
}
\details{
User specifies the settings to create a default config for an estimation module
}
