% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createModuleConfig}
\alias{createModuleConfig}
\title{createModuleConfig}
\usage{
createModuleConfig(
  moduleId = "about",
  tabName = "About",
  shinyModulePackage = "OhdsiShinyModules",
  moduleUiFunction = "aboutViewer",
  moduleServerFunction = "aboutServer",
  moduleDatabaseConnectionKeyService = "resultDatabaseDetails",
  moduleDatabaseConnectionKeyUsername = NULL,
  moduleInfoBoxFile = "aboutHelperFile()",
  moduleIcon = "info",
  resultDatabaseDetails,
  useKeyring = TRUE
)
}
\arguments{
\item{moduleId}{The shiny id for the tab containing the module UI}

\item{tabName}{The name of the tab in the shiny app (this will be the side menu button text)}

\item{shinyModulePackage}{The R package to find the server and UI functions}

\item{moduleUiFunction}{The name of the UI function in the R package shinyModulePackage}

\item{moduleServerFunction}{The name of the server function in the R package shinyModulePackage. If null package will try to try to load ui, helper and moduleServer from local session}

\item{moduleDatabaseConnectionKeyService}{The keyring service or the system environment variable with the result database details}

\item{moduleDatabaseConnectionKeyUsername}{The keyring username or the system environment variable with the result database details}

\item{moduleInfoBoxFile}{The function in the R package shinyModulePackage that contains info text}

\item{moduleIcon}{The icon to use for the side menu button}

\item{resultDatabaseDetails}{A list containing the result database schema (schema), the tablePrefix for the results and other optional settings}

\item{useKeyring}{whether to save the resultDatabaseDetails to a keyring or system environmental variable}
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the config specification
}
\details{
User specifies the settings to create a config for a module
}
