% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewShiny.R
\name{createShinyApp}
\alias{createShinyApp}
\title{createShinyApp}
\usage{
createShinyApp(
  config,
  connection = NULL,
  connectionDetails = NULL,
  usePooledConnection = TRUE
)
}
\arguments{
\item{config}{The json with the app config}

\item{connection}{A connection to the results}

\item{connectionDetails}{A DatabaseConnector::connectionDetails connection to the results database}

\item{usePooledConnection}{use a pooled database connection or not - set to true for multi-user environments (default)}
}
\value{
Shiny app instance
}
\description{
Create a shiny app for a shiny server
}
\details{
User specifies the json config and connection
}
