% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewShiny.R
\name{createShinyApp}
\alias{createShinyApp}
\title{createShinyApp}
\usage{
createShinyApp(config, connection)
}
\arguments{
\item{config}{The json with the app config}

\item{connection}{A connection to the results}
}
\value{
The shiny app will open
}
\description{
Create a shiny app for a shiny server
}
\details{
User specifies the json config and connection
}
