% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewShiny.R
\name{createShinyApp}
\alias{createShinyApp}
\title{createShinyApp}
\usage{
createShinyApp(
  config,
  connection,
  resultDatabaseSettings = list(schema = "main", cgTablePrefix = "cg_", cdTablePrefix =
    "cd_", cTablePrefix = "c_", plpTablePrefix = "plp_", pvTablePrefix = "pv_",
    cmTablePrefix = "cm_", sccsTablePrefix = "sccs_", esTablePrefix = "es_",
    incidenceTablePrefix = "i_", databaseTable = "database_meta_table",
    databaseTablePrefix = "")
)
}
\arguments{
\item{config}{The json with the app config}

\item{connection}{A connection to the results}

\item{resultDatabaseSettings}{A list with the result schema and table prefixes}
}
\value{
The shiny app will open
}
\description{
Create a shiny app for a shiny server
}
\details{
User specifies the json config and connection
}
