% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewShiny.R
\name{viewShiny}
\alias{viewShiny}
\title{viewShiny}
\usage{
viewShiny(
  config,
  connection,
  resultDatabaseSettings = createDefaultResultDatabaseSettings(),
  connectionDetails = NULL,
  usePooledConnection = TRUE,
  studyDescription = NULL,
  title = "OHDSI Analysis Viewer",
  protocolLink = "http://ohdsi.org"
)
}
\arguments{
\item{config}{The json with the app config}

\item{connection}{A connection to the results}

\item{resultDatabaseSettings}{A list with the result schema and table prefixes}

\item{connectionDetails}{A DatabaseConnector::connectionDetails connection to the results database}

\item{usePooledConnection}{Use a pooled database connection or not - set to true for multi-user environments (default)}

\item{studyDescription}{A human-readable character string describing the study/analysis}
  
\item{title}{The title for the app.  Defaults to: OHDSI Analysis Viewer}

\item{protocolLink}{A link to a site containing the study protocol}

}
\value{
The shiny app will open
}
\description{
Open the shiny app
}
\details{
User specifies the json config and connection
}
